# Conflicts:
#	src/jsc/VirtualMachine.rs
#	src/runtime/bake/DevServer.rs
#	src/runtime/cli/install_completions_command.rs
#	src/runtime/test_runner/snapshot.rs

+++ src/jsc/bindings/JSBuffer.cpp
@@ -151,7 +151,7 @@ namespace Bun {
 
 // Use a JSString* here to avoid unnecessarily joining the rope string.
 // If we're only getting the length property, it won't join the rope string.
-std::optional<double> byteLength(JSC::JSString* str, JSC::JSGlobalObject* lexicalGlobalObject, WebCore::BufferEncodingType encoding)
+std::optional<size_t> byteLength(JSC::JSString* str, JSC::JSGlobalObject* lexicalGlobalObject, WebCore::BufferEncodingType encoding)
 {
     if (str->length() == 0)
         return 0;
@@ -197,7 +197,7 @@ std::optional<double> byteLength(JSC::JSString* str, JSC::JSGlobalObject* lexica
         }
 
         // https://github.com/nodejs/node/blob/e676942f814915b2d24fc899bb42dc71ae6c8226/lib/buffer.js#L579
-        return static_cast<double>((length * 3) >> 2);
+        return static_cast<size_t>((length * 3) >> 2);
     }
 
     case WebCore::BufferEncodingType::hex: {
